+++ client/infinity-game/src/router.js
@@ -2,6 +2,7 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import Home from './views/Home.vue'
 import Cart from './views/cart.vue'
+import Payment from './views/paymentSummary.vue'
 
 Vue.use(Router)
 
@@ -17,7 +18,14 @@ export default new Router({
     {
       path : '/cart',
       name : 'cart',
-      component : Cart,       
+      component : Cart,
+      children : [
+        {
+          path: 'payment',
+          name : 'payment',
+          component : Payment
+        }
+      ]       
     }
   ]
 })
